+++ src/main/scala/org/mmisw/orr/ont/service/TripleStoreService.scala
@@ -1,5 +1,7 @@
 package org.mmisw.orr.ont.service
 
+case class TermResponse(contents: String, contentTye: String)
+
 trait TripleStoreService {
 
   def setFormats(formats: Map[String, String]): Unit
@@ -40,5 +42,16 @@ trait TripleStoreService {
    * Clears the triple store.
    */
   def unloadAll(): Either[Throwable, String]
+
+  /**
+    * Resolves a URI via SPARQL query to retrieve associated properties.
+    *
+    * @param uri           Requested URI
+    * @param formatOpt     Desired format, which takes precedence if given
+    * @param acceptHeader  Accept header when format is not given
+    * @return (result, contentType)
+    */
+  def resolveTermUri(uri: String, formatOpt: Option[String] = None, acceptHeader: List[String] = Nil
+                    ): Either[Error, TermResponse]
 
 }