+++ ChangeLog.md
@@ -1,5 +1,26 @@
 ## change log ##
 
+* 2016-06-16: 0.3.2:
+  - \#14 "dispatch term request"
+    - based on SPARQL query for the properties of the term.
+    - added agraph.sparqlEndpoint config property to capture the complete URL of the SPARQL endpoint. 
+      Although most of the associated config pieces are already captured in other properties, the 
+      associated request without protocol (eg. http:) causes AG to respond with a 301 redirect.
+      Although Dispatch could be configured to follow redirects, decided to captured the final location
+      to avoid the extra round trip.
+    - As in previous system, the term query will be either a "select" or a "construct" depending on the desired format. 
+      This format can be specified via explicit "format" parameter:
+        `http 'http://mmisw.org/ont/api/v0/ont?uri=https://mmisw.org/ont/test/avoc/term1&format=json'`
+      or via accept header:
+        `http 'http://mmisw.org/ont/api/v0/ont?uri=https://mmisw.org/ont/test/avoc/term1' Accept:text/rdf+n3`
+      There's an ad hoc mapping from the format to corresponding mime-type for the request
+    - GET /ont: parameters considered and dispatched with the following precedence:
+      - "uri": tries ontology first, then term. In this case, 404 will never occur as the term attempt will always 
+        succeed (unless there's an unexpected exception) because the sparql query can have an "empty" response   
+      - "ouri": only tries ontology, with 404 response if not found 
+      - "turi": performs term query, no 404 can occur as already explained
+    - TODO SelfHostedController
+    
 * 2016-06-11: 0.3.2:
   - \#10: "capture visibility attribute for ontologies"
     - for convenience at the moment, optional in OntologyVersion model (but to be required, and with no default value)